docs: filled out README.md, and reorganized internally
