+++ src/index.ts
@@ -1,9 +1,11 @@
+export * from "./comments.js";
 export * from "./compilerOptions.js";
 export * from "./flags.js";
-export * from "./forEachComment.js";
-export * from "./forEachToken.js";
-export * from "./scope.js";
-export * from "./typeguards.js";
-export * from "./types/get.js";
-export * from "./types/has.js";
-export * from "./types/is.js";
+export * from "./modifiers.js";
+export * from "./nodes/typeGuards.js";
+export * from "./scopes.js";
+export * from "./syntax.js";
+export * from "./tokens.js";
+export * from "./types/getters.js";
+export * from "./types/typeGuards.js";
+export * from "./types/utilities.js";
