+++ src/scopes.ts
